fix(Segments Looping) : Refactor code and create segmentUtils to handleOnSegmentChange for SEG and RTS (#4614)
