+++ extensions/cornerstone-dicom-rt/src/viewports/OHIFCornerstoneRTViewport.tsx
@@ -2,6 +2,7 @@ import React, { useCallback, useEffect, useRef, useState } from 'react';
 import PropTypes from 'prop-types';
 import { ViewportActionArrows } from '@ohif/ui';
 import { useViewportGrid } from '@ohif/ui-next';
+import { onSegmentChange as handleSegmentChange } from '../../../cornerstone/src/utils/segmentUtils';
 
 import promptHydrateRT from '../utils/promptHydrateRT';
 import _getStatusComponent from './_getStatusComponent';
@@ -141,38 +142,9 @@ function OHIFCornerstoneRTViewport(props: withAppTypes) {
 
   const onSegmentChange = useCallback(
     direction => {
-      const segmentationId = rtDisplaySet.displaySetInstanceUID;
-      const segmentation = segmentationService.getSegmentation(segmentationId);
-
-      const { segments } = segmentation;
-
-      const numberOfSegments = Object.keys(segments).length;
-      //Get activeSegment each time because the user can select any segment from the list and thus the index should be updated
-      const activeSegment = segmentationService.getActiveSegment(viewportId);
-      if (activeSegment) {
-        const activeSegmentIndex = Object.values(segments).findIndex(
-          segment => segment.segmentIndex === activeSegment.segmentIndex
-        );
-        //from the activeSegment get the actual obeject array index to be used
-        selectedSegmentObjectIndex = activeSegmentIndex;
-      }
-      let newSelectedSegmentIndex = selectedSegmentObjectIndex + direction;
-
-      //Handle looping through list of segments
-      if (newSelectedSegmentIndex > numberOfSegments - 1) {
-        newSelectedSegmentIndex = 0;
-      } else if (newSelectedSegmentIndex < 0) {
-        newSelectedSegmentIndex = numberOfSegments - 1;
-      }
-
-      //convert segmentationId from object array index to property value of type Segment
-      //Functions below uses the segmentIndex object attribute so we have to do the conversion
-      const keyIndex = Object.values(segments)[newSelectedSegmentIndex]?.segmentIndex;
-      segmentationService.setActiveSegment(segmentationId, keyIndex);
-      segmentationService.jumpToSegmentCenter(segmentationId, keyIndex, viewportId);
-      selectedSegmentObjectIndex = newSelectedSegmentIndex;
+      handleSegmentChange(direction, rtDisplaySet, viewportId, selectedSegmentObjectIndex);
     },
-    [selectedSegmentObjectIndex, segmentationService]
+    [selectedSegmentObjectIndex]
   );
 
   useEffect(() => {
