+++ extensions/cornerstone-dicom-seg/src/viewports/OHIFCornerstoneSEGViewport.tsx
@@ -7,6 +7,7 @@ import promptHydrateSEG from '../utils/promptHydrateSEG';
 import _getStatusComponent from './_getStatusComponent';
 import { usePositionPresentationStore } from '@ohif/extension-cornerstone';
 import { SegmentationRepresentations } from '@cornerstonejs/tools/enums';
+import { onSegmentChange as handleSegmentChange } from '../../../cornerstone/src/utils/segmentUtils';
 
 const SEG_TOOLGROUP_BASE_NAME = 'SEGToolGroup';
 
@@ -131,38 +132,7 @@ function OHIFCornerstoneSEGViewport(props: withAppTypes) {
 
   const onSegmentChange = useCallback(
     direction => {
-      const segmentationId = segDisplaySet.displaySetInstanceUID;
-      const segmentation = segmentationService.getSegmentation(segmentationId);
-
-      const { segments } = segmentation;
-
-      const numberOfSegments = Object.keys(segments).length;
-
-      //Get activeSegment each time because the user can select any segment from the list and thus the index should be updated
-      const activeSegment = segmentationService.getActiveSegment(viewportId);
-      if (activeSegment) {
-        const activeSegmentIndex = Object.values(segments).findIndex(
-          segment => segment.segmentIndex === activeSegment.segmentIndex
-        );
-        //from the activeSegment get the actual obeject array index to be used
-        selectedSegmentObjectIndex = activeSegmentIndex;
-      }
-      let newSelectedSegmentIndex = selectedSegmentObjectIndex + direction;
-
-      //Handle looping through list of segments
-      if (newSelectedSegmentIndex > numberOfSegments - 1) {
-        newSelectedSegmentIndex = 0;
-      } else if (newSelectedSegmentIndex < 0) {
-        newSelectedSegmentIndex = numberOfSegments - 1;
-      }
-
-      //convert segmentationId from object array index to property value of type Segment
-      //Functions below uses the segmentIndex object attribute so we have to do the conversion
-      const segmentIndex = Object.values(segments)[newSelectedSegmentIndex]?.segmentIndex;
-
-      segmentationService.setActiveSegment(segmentationId, segmentIndex);
-      segmentationService.jumpToSegmentCenter(segmentationId, segmentIndex, viewportId);
-      selectedSegmentObjectIndex = newSelectedSegmentIndex;
+      handleSegmentChange(direction, segDisplaySet, viewportId, selectedSegmentObjectIndex);
     },
     [selectedSegmentObjectIndex]
   );
